sha256.h: Changed message len size to 64 bit uint to fit size

+++ sha256/sha256.h
@@ -20,7 +20,7 @@ public:
 private:
     void to_binary(const std::string &str);
     void pad();
-    uint8_t get_m_size();
+    uint64_t get_m_size();
 };
 
 #endif
